fix(utils): set type for convertFunction args
