+++ packages/components/form/src/components/fields/text/text.tsx
@@ -6,11 +6,11 @@ import useForms from '../../../hooks/useForms/useForms';
 
 const Text: FC<TextProps> = ({ formId, ...textFieldProps }) => {
   const { forms } = useForms();
+
   const formMethod = forms?.[formId];
   const { field } = useController({
     name: textFieldProps.id,
     control: formMethod.control,
-    rules: { required: true },
   });
 
   return <TextField {...field} {...textFieldProps} value={field.value ?? ''} />;
