+++ packages/utils/src/convertFunctions/convertFunction.ts
@@ -1,14 +1,17 @@
 import { TConvertFunctionProps } from './convertFunction.types';
 
-const convertFunction = (fn: TConvertFunctionProps, ...props: string[]) => {
+const convertFunction = <T = unknown>(
+  fn: TConvertFunctionProps,
+  ...props: string[]
+): T => {
   if (!fn)
-    return () => {
+    return (() => {
       return {};
-    };
+    }) as T;
 
-  if (typeof fn === 'function') return fn;
+  if (typeof fn === 'function') return fn as T;
 
-  return new Function(...props, fn);
+  return new Function(...props, fn) as T;
 };
 
 export default convertFunction;
